# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Astral Modulator Public API
  version: 2.0.0-scaffold
  license:
    name: MIT
    identifier: MIT
  description: |
    人类与多 LLM Agent 的协作控制面。CLI（astral-cli 仓库）与 Web GUI 共用本协议。

    v2（容器化任务树，TODO.md D15）：workspace 与 task 都是容器，子任务集合
    统一为 GET/POST <container>/children；全量平铺列表与 /tasks/search 已移除，
    平面查询走 /workspaces/{id}/task-search。protocol_version 升为 2，
    与 v1 不兼容（破坏性变更，开发期不做兼容窗口）。

    已裁决的文档分歧（详见 TODO.md「文档分歧裁决」D1-D4）：
    - Device flow 采用 `/auth/device/authorizations*` 路径。
    - 当前身份端点为 `/auth/me`。

    通用约定：
    - 所有时间戳为 RFC3339 UTC。
    - ID 为 `prefix_uuidv7` 形式的不透明字符串，客户端不得解析内部结构。
    - 错误响应统一使用 ErrorEnvelope；CLI 依赖稳定 error.code。
    - 乐观并发统一使用 body 内 expected_revision（protocol §6 方案 A）。
servers:
  - url: /api/v1
    description: 路径前缀固定 /api/v1（不随协议大版本变化）；协议大版本以
      /.well-known/astral 的 protocol_version 为准
...
security:
  - bearerAuth: []
paths:
...
  /workspaces/{workspace_id}/tags:
    get:
      tags: [tags]
      summary: workspace tag 列表（proposal 第一步响应的 existing_tags 来源）
      operationId: listTags
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
      responses:
        "200":
          description: tag 分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagPage"
  /workspaces/{workspace_id}/tag-proposals:
    post:
      tags: [tags]
      summary: Tag 两步确认第一步（不创建正式 tag；支持 Idempotency-Key）
      operationId: proposeTag
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TagProposalRequest" }
      responses:
        "201":
          description: proposal 已创建（含全量 existing_tags 供比对）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TagProposal" }
  /tag-proposals/{proposal_id}/confirm:
    post:
      tags: [tags]
      summary: Tag 两步确认第二步（CLI 固定拼写 --confirm）
      operationId: confirmTagProposal
      parameters:
        - { name: proposal_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm_code, name]
              properties:
                confirm_code: { type: string, description: 服务端下发的校验码（TTL 内单次有效） }
                name: { type: string, description: 必须与 proposal 的 canonical 输入一致 }
      responses:
        "200":
          description: 确认成功（create/rename/delete 已生效）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tag" }
        "409": { $ref: "#/components/responses/Error" }   # TAG_PROPOSAL_EXPIRED / TAG_ALREADY_EXISTS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Human access token 或 Agent/Service credential（不共用）
    cookieAuth:
      type: apiKey
      in: cookie
      name: astral_session
      description: Web GUI HttpOnly+Secure+SameSite session（与 CLI Bearer 同 actor 不同 session）
  parameters:
    WorkspaceId:
      name: workspace_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
...
  responses:
    Error:
      description: 统一错误 envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    Id:
      type: string
      pattern: "^[a-z]{2,3}_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      description: 不透明 ID：prefix_uuidv7。前缀 usr/agt/svc/srv/ws/tsk/tgp/tag/doc/cfl/prs/msg/evt/cred/ses/dev/req/aud/apv/inv。
      examples: ["ws_0192ab34-56cd-7ef8-9a01-234567890abc"]
...
    ErrorCode:
      type: string
      enum:
        - AUTH_REQUIRED
        - TOKEN_EXPIRED
        - TOKEN_REVOKED
        - INSUFFICIENT_SCOPE
        - SERVER_NOT_FOUND
        - WORKSPACE_NOT_FOUND
        - WORKSPACE_ALREADY_BOUND
        - WORKSPACE_NAME_TAKEN
        - TASK_NOT_FOUND
        - TASK_ALREADY_CLAIMED
        - TASK_LEASE_EXPIRED
        - TAG_PROPOSAL_EXPIRED
        - APPROVAL_EXPIRED        # approval 过期或已裁决（409）
        - INVITE_INVALID          # 邀请码不存在/已兑换/已撤销/已过期统一此码（防探测，400）
        - EMAIL_TAKEN             # email 已被注册（409，仅邀请码验证通过后可达）
        - TAG_ALREADY_EXISTS
        - REVISION_CONFLICT
        - DOCUMENT_CONFLICT
        - RATE_LIMITED
        - CLIENT_VERSION_UNSUPPORTED
        - VALIDATION_FAILED
        - NOT_FOUND               # 未知路由 / 无专用码的次级资源不存在（HTTP 404）
        - AUTHORIZATION_PENDING   # device flow 轮询（A1/RFC 8628），HTTP 400 + retryable
        - SLOW_DOWN               # device flow 轮询过快（A1），HTTP 400 + retryable
        - INTERNAL_ERROR
        - NOT_IMPLEMENTED   # 仅脚手架阶段（HTTP 501），正式端点不得返回
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, retryable, request_id]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message: { type: string }
            retryable: { type: boolean }
            details: { type: object, additionalProperties: true }
            request_id: { type: string }
...
    Tag:
      type: object
      required: [id, workspace_id, name]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        workspace_id: { $ref: "#/components/schemas/Id" }
        name: { type: string }
    TagProposalRequest:
      type: object
      required: [action, name]
      properties:
        action: { type: string, enum: [create, rename, delete] }
        name: { type: string, minLength: 1, maxLength: 64 }
        target_tag_id: { $ref: "#/components/schemas/Id" }
    TagProposal:
      type: object
      required: [proposal_id, confirm_code, expires_at, existing_tags]
      properties:
        proposal_id: { $ref: "#/components/schemas/Id" }
        confirm_code: { type: string, description: 短 TTL 校验码；confirm 时原样提交 }
        expires_at: { type: string, format: date-time }
        existing_tags:
          type: array
          items: { $ref: "#/components/schemas/Tag" }
          description: 全量已有 tag，供 Agent 比对后三思
...
    PageMeta:
      type: object
      required: [next_cursor]
      properties:
        next_cursor:
          type: [string, "null"]
          description: null 表示没有更多
...
    TagPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/Tag" } }
...